Rename decorator files to inspector files
