+++ lib/ruby_jard/inspectors/array_inspector.rb
