+++ lib/ruby_jard/inspectors/attributes_inspector.rb
