+++ lib/ruby_jard/inspectors/base.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
 require 'ruby_jard/inspectors/primitive_inspector'
-require 'ruby_jard/inspectors/array_decorator'
-require 'ruby_jard/inspectors/string_decorator'
-require 'ruby_jard/inspectors/hash_decorator'
-require 'ruby_jard/inspectors/struct_decorator'
-require 'ruby_jard/inspectors/object_decorator'
-require 'ruby_jard/inspectors/attributes_decorator'
-require 'ruby_jard/inspectors/rails_decorator'
+require 'ruby_jard/inspectors/array_inspector'
+require 'ruby_jard/inspectors/string_inspector'
+require 'ruby_jard/inspectors/hash_inspector'
+require 'ruby_jard/inspectors/struct_inspector'
+require 'ruby_jard/inspectors/object_inspector'
+require 'ruby_jard/inspectors/attributes_inspector'
+require 'ruby_jard/inspectors/rails_inspectors'
 
 module RubyJard
   module Inpsectors
