+++ lib/ruby_jard/inspectors/hash_inspector.rb
