+++ lib/ruby_jard/inspectors/object_inspector.rb
