+++ lib/ruby_jard/inspectors/rails_inspectors.rb
