+++ lib/ruby_jard/inspectors/string_inspector.rb
