+++ lib/ruby_jard/inspectors/struct_inspector.rb
